==> notebooks/00_exploration.ipynb <==
DATA_URL = r'https://cdn.stackoverflow.co/files/jo7n4k8s/production/49915bfd46d0902c3564fd9a06b509d08a20488c.zip/stack-overflow-developer-survey-2023.zip'
DATA_PATH = '../data/raw/'
SURVEY_FILENAME = 'survey_results_public.csv'

import pandas as pd 
import urllib.request
import zipfile
import os

pd.options.display.max_rows = 10000

### Functions

def download_and_unzip(url, extract_path):

  zip_name = url.split("/")[-1]
  with urllib.request.urlopen(url) as response:
    with open(f"{extract_path}/{zip_name}", "wb") as file:
      file.write(response.read())

  with zipfile.ZipFile(f"{extract_path}/{zip_name}", "r") as zip_file:
    zip_file.extractall(extract_path)

  return zip_name

## Explore

zip_name = download_and_unzip(DATA_URL, DATA_PATH)

results_path = os.path.join(DATA_PATH, SURVEY_FILENAME)

# Read data and print shape
raw_df = pd.read_csv(results_path)
raw_df.shape

# Display random answer 
# Observations: Multiple answers need to be splitted 
# Reference to the schema needed to understand
raw_df.sample(1).iloc[0]

# Print the general information of the data frame 
raw_df.info()

# Get stats for the numerical column
raw_df.describe()

# Investigate the questionable objects columns
questionable_cols = ['YearsCodePro', 'YearsCode']

for col in questionable_cols: 
    print(col)
    print(raw_df[col].unique().tolist())
    print('--------------------------')
    print()
